Adapt to SONATA standard  * Change report data datatype from double to float

+++ src/reports/data/sonata_data.cpp
@@ -54,7 +54,7 @@ void SonataData::prepare_buffer(size_t max_buffer_size) {
 
     // Calculate the timesteps that fit given the buffer size
     {
-        uint32_t max_steps_to_write = max_buffer_size / (sizeof(double) * total_elements_);
+        uint32_t max_steps_to_write = max_buffer_size / (sizeof(float) * total_elements_);
         uint32_t common_max_steps_to_write =
             Implementation::get_max_steps_to_write(report_name_, max_steps_to_write);
         if (common_max_steps_to_write < num_steps_) {  // More steps asked that buffer can contain
